requirements.txt
Django==4.2.7
djangorestframework
django-rest-framework-token